requirements.txt
pydantic>=2.0
pytest
sentence-transformers
pandas
pyyaml
httpx
openai
rich
python-dotenv
tiktoken
